refactor struct, getting the prune out of eval.rs
